🔨 Refactor: 이메일 전송 로직 QueryService에서 구현하도록 수정
