+++ src/main/java/qastudio/backend/domain/project/service/TeamMemberCommandService.java
@@ -12,7 +12,6 @@ import qastudio.backend.domain.project.entity.UserProject;
 import java.util.List;
 
 public interface TeamMemberCommandService {
-    void inviteMembers(Long projectId, List<TeamMemberRequest.MemberEmail> memberEmailList);
     void deleteMembers(Long projectId, TeamMemberRequest.MemberEmail deleteMember);
     Long inviteMember(String token);
 }
