+++ src/main/java/qastudio/backend/domain/project/service/TeamMemberQueryService.java
@@ -15,4 +15,6 @@ public interface TeamMemberQueryService {
     List<AccountTable> searchMember(Long projectId, String email);
 
     List<UserProject> getTeamMemberExceptLeader(Long projectId);
+
+    void inviteMembers(Long projectId, List<TeamMemberRequest.MemberEmail> memberEmailList);
 }
